Add tests

+++ src/main/kotlin/com/x0/hatonekoe/service/Constant.kt
@@ -6,5 +6,5 @@ object Constant {
     val WINDOW_MIN_HEIGHT: Double = 150.0
     val TEXTAREA_FONT_SIZE: Double = 16.0
 
-    val OPEN_FILE_SIZE_LIMIT_BYTE: Int = 10_000_000 // 10MB
+    val OPEN_FILE_SIZE_LIMIT_BYTE: Long = 10_000_000 // 10MB
 }
